+++ apps/explorer/config/prod.exs
@@ -16,6 +16,10 @@ config :explorer, Explorer.Repo.Account,
   prepare: :unnamed,
   timeout: :timer.seconds(60)
 
+config :explorer, Explorer.Repo.PolygonEdge,
+  prepare: :unnamed,
+  timeout: :timer.seconds(60)
+
 config :explorer, Explorer.Tracer, env: "production", disabled?: true
 
 config :logger, :explorer,