+++ apps/explorer/lib/explorer/chain.ex
@@ -2232,7 +2232,7 @@ defmodule Explorer.Chain do
       )
 
     query
-    |> add_fetcher_limit(limited?)
+    |> add_coin_balances_fetcher_limit(limited?)
     |> Repo.stream_reduce(initial, reducer)
   end
 
@@ -6315,6 +6315,14 @@ defmodule Explorer.Chain do
     limit(query, ^token_balances_fetcher_limit)
   end
 
+  defp add_coin_balances_fetcher_limit(query, false), do: query
+
+  defp add_coin_balances_fetcher_limit(query, true) do
+    coin_balances_fetcher_limit = Application.get_env(:indexer, :coin_balances_fetcher_init_limit)
+
+    limit(query, ^coin_balances_fetcher_limit)
+  end
+
   def put_has_token_transfers_to_tx(query, true), do: query
 
   def put_has_token_transfers_to_tx(query, false) do
@@ -6341,5 +6349,10 @@ defmodule Explorer.Chain do
       )
 
     Repo.all(query)
+  end
+
+  @spec default_paging_options() :: map()
+  def default_paging_options do
+    @default_paging_options
   end
 end